requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cryptography==1.1
python-dotenv==1.0.1
cryptography==44.0.0
bcrypt==4.0.1
gunicorn==21.2.0
django-extensions==3.2.3
pyotp==2.9.0
django-cors-headers==4.1.0
